requirements.txt
Cerberus==1.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-modal-forms==1.5.0
django-bootstrap4==0.0.8
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-toolbelt==0.0.1
django-widget-tweaks==1.4.5
gunicorn==19.9.0
psycopg2==2.8.2
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
whitenoise==4.1.2
